feat: Change search fields end-point.
