+++ src/api/ADempiere/fields/binary.js
@@ -1,0 +1,54 @@
+/**
+ * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ * Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Contributor(s): Elsio Sanchez elsiosanches@gmail.com https://github.com/elsiosanchez
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Get entity with binary by identifier
+ * @param {string} tableName
+ * @param {string} recordUuid
+ */
+export function getResource({
+  uuid,
+  tableName
+}) {
+  const { getEntity } = require('@/api/ADempiere/common/persistence.js')
+
+  return getEntity({
+    recordUuid: uuid,
+    tableName
+  })
+}
+
+/**
+ * Update an existing binary by id or uuid
+ * @param {string} tableName
+ * @param {string} recordUuid
+ * @param {object} binaryFile
+ */
+export function updateResource({
+  uuid,
+  tableName,
+  binaryFile
+}) {
+  const { updateEntity } = require('@/api/ADempiere/userInterface/entities.ts')
+
+  return updateEntity({
+    recordUuid: uuid,
+    tableName,
+    attributesList: binaryFile
+  })
+}
