+++ src/api/ADempiere/fields/lookups.ts
