+++ src/api/ADempiere/fields/search/product.ts
