+++ src/components/ADempiere/FieldDefinition/FieldProductAttribute/detailAttributeInstance.vue
@@ -1,6 +1,6 @@
 <!--
  ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
- Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
  Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
  This program is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -131,7 +131,7 @@ import {
   requestGetProductAttributeSet,
   requestGetProductAttributeSetInstace,
   requestSaveAttributeSetInstance
-} from '@/api/ADempiere/field/productAttribute'
+} from '@/api/ADempiere/fields/productAttribute.js'
 
 // Constants
 import {
