+++ src/components/ADempiere/Form/VPOS/BusinessPartner/shippingFieldLocation/locationAddressForm.vue
@@ -75,7 +75,7 @@ import { DISPLAY_COLUMN_PREFIX } from '@/utils/ADempiere/dictionaryUtils'
 import {
   createLocationAddress,
   updateLocationAddress
-} from '@/api/ADempiere/field/location.js'
+} from '@/api/ADempiere/fields/location.js'
 
 // Utils and Helpers Methods
 import { showNotification } from '@/utils/ADempiere/notification.js'
